requirements.txt
pandas
sqlalchemy
pyarrow
gdown
python-dotenv
